requirements-vercel.txt
fastapi
uvicorn
langchain-core
langchain-openai
chromadb
openai
pypdf
python-dotenv
